Backport 907f9ee7083df384a7433cb933522680e70dcb1f
