+++ jdk/test/tools/launcher/VersionCheck.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2007, 2014, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2007, 2020, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -31,11 +31,12 @@
  */
 
 import java.io.File;
-import java.io.FileFilter;
-import java.util.Map;
 import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Map;
+import java.util.Set;
 
 public class VersionCheck extends TestHelper {
 
@@ -225,34 +226,6 @@ public class VersionCheck extends TestHelper {
             System.out.println("All Version string comparisons: PASS");
         } else {
             throw new AssertionError("Some tests failed");
-        }
-    }
-
-    static class ToolFilter implements FileFilter {
-        final Iterable<String> exclude ;
-        protected ToolFilter(String... exclude) {
-            List<String> tlist = new ArrayList<>();
-            this.exclude = tlist;
-            for (String x : exclude) {
-                String str = x + ((isWindows) ? EXE_FILE_EXT : "");
-                tlist.add(str.toLowerCase());
-            }
-        }
-        @Override
-        public boolean accept(File pathname) {
-            if (!pathname.isFile() || !pathname.canExecute()) {
-                return false;
-            }
-            String name = pathname.getName().toLowerCase();
-            if (isWindows && !name.endsWith(EXE_FILE_EXT)) {
-                return false;
-            }
-            for (String x : exclude) {
-                if (name.endsWith(x)) {
-                    return false;
-                }
-            }
-            return true;
         }
     }
 }
